rag-prototype/requirements.txt
google-adk
vertexai
google-cloud-aiplatform
google-cloud-storage
google-cloud-pubsub
functions-framework
cloudevents
fastapi
uvicorn
gradio
python-dotenv
requests
